+++ Rock/Configuration/AttributeValueConfig.cs
@@ -16,7 +16,7 @@
 //
 using System.Configuration;
 
-namespace RockWeb
+namespace Rock.Configuration
 {
     /// <summary>
     /// Read Attribute Value node in rockConffig
