+++ RockWeb/web.config
@@ -3,7 +3,7 @@
   <configSections>
     <section name="resizer" type="ImageResizer.ResizerSection,ImageResizer" requirePermission="false" />
     <section name="quartz" type="System.Configuration.NameValueSectionHandler, System, Version=1.0.5000.0,Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-    <section name="rockConfig" type="RockWeb.RockConfig" />
+    <section name="rockConfig" type="Rock.Configuration.RockConfig, Rock" />
   </configSections>
   <quartz>
     <add key="quartz.scheduler.instanceName" value="RockSchedulerIIS" />
@@ -20,7 +20,7 @@
   </resizer>
   <rockConfig>
     <attributeValues>
-      <attributeValue attributeKey="NodeUrl" entityTypeId="406" entityId="16" value="value"/>
+      <!--<attributeValue attributeKey="NodeUrl" entityTypeId="406" entityId="16" value="value"/>-->
     </attributeValues>
   </rockConfig>
   <connectionStrings configSource="web.ConnectionStrings.config" />
